requirements.txt
apex==0.9.10dev
effdet==0.4.1
grad_cam==1.4.8
matplotlib==3.5.2
numba==0.55.1
numpy==1.21.5
Pillow==9.5.0
pycocotools==2.0.6
scikit_learn==1.0.2
timm==0.9.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
wandb==0.13.9